Version 0.4.2
Bug fix: Room qualities automatically set to zero due to casting error.
Error in Tester used to test this method also fixed.

New solve method for MinUpgradesSTSatIP solver. Takes minimum and
average satisfaction constraints as input to reduce IP repetition for
repeated simulations.

+++ Hotel Simulation (Single Day)/src/com/henryrobbins/hotel/Instance.java
@@ -144,7 +144,7 @@ public final class Instance {
 		HashMap<Room, Double> quality= new HashMap<>();
 		for (Room room : rooms) {
 			int feasibleGuests= 0;
-			int totalQuality= 0;
+			double totalQuality= 0;
 			for (Guest guest : guests) {
 				if (room.type() >= guest.type()) {
 					feasibleGuests++ ;
